pinches esquemas de sql

+++ WebApi.Model/VentaEntities.cs
@@ -16,5 +16,5 @@ public class DetalleVenta{
     public int IdProducto { get; set; } 
     public int Cantidad { get; set; }
     public decimal PrecioUnitario { get; set; }
-    public int Subtotal { get; set; } 
+    public decimal Subtotal { get; set; } 
 }
